Minor fixes
- Allow dev tool toggle when running in DEV mode
- Minor debatable grammer on Index page
- Check process exists before killing (was throwing error when app quit)

+++ DervetApp/src/main/service/dervet.js
@@ -71,8 +71,10 @@ const listenForExit = pythonProcess => (
 );
 
 export const exitPythonProcess = (pythonProcess) => {
-  pythonProcess.kill();
-  pythonProcess = null;
+  if (pythonProcess) {
+    pythonProcess.kill();
+    pythonProcess = null;
+  }
 };
 
 export const callDervet = (modelParametersPath) => {
